+++ src/utils/mod.zig
@@ -1,2 +1,1 @@
-pub const waybar = @import("waybar.zig");
 pub const config = @import("config.zig");